metadata is now less redundant
