+++ data/general.json
@@ -179,81 +179,7 @@
                 "on": "12/11/2024 20:39",
                 "off": "12/11/2024 21:54"
             }
-        ],
-        "lsc_sample_times": {
-            "IV": {
-                "2-0-x": {
-                    "planned": "12/10/2024 10:51",
-                    "actual": "12/10/2024 10:32"
-                },
-                "2-1-x": {
-                    "planned": "12/11/2024 10:51",
-                    "actual": "12/11/2024 9:15"
-                },
-                "2-2-x": {
-                    "planned": "12/12/2024 10:51",
-                    "actual": "12/12/2024 9:00"
-                },
-                "2-3-x": {
-                    "planned": "12/14/2024 10:51",
-                    "actual": "12/14/2024 12:58"
-                },
-                "2-4-x": {
-                    "planned": "12/16/2024 10:51",
-                    "actual": "12/16/2024 12:49"
-                },
-                "2-5-x": {
-                    "planned": "12/19/2024 10:51",
-                    "actual": null
-                },
-                "2-6-x": {
-                    "planned": "12/22/2024 10:51",
-                    "actual": null
-                },
-                "2-7-x": {
-                    "planned": "12/27/2024 10:51",
-                    "actual": null
-                },
-                "2-8-x": {
-                    "planned": "1/1/2025 10:51",
-                    "actual": null
-                },
-                "2-9-x": {
-                    "planned": "1/6/2025 10:51",
-                    "actual": null
-                }
-            },
-            "OV": {
-                "2-0-x": {
-                    "planned": "11/4/2024 9:30",
-                    "actual": "12/10/2024 10:32"
-                },
-                "2-1-x": {
-                    "planned": "12/19/2024 10:51",
-                    "actual": null
-                },
-                "2-2-x": {
-                    "planned": "1/14/2025 10:51",
-                    "actual": null
-                },
-                "2-3-x": {
-                    "planned": "12/10/2024 10:51",
-                    "actual": null
-                },
-                "2-4-x": {
-                    "planned": "12/10/2024 10:51",
-                    "actual": null
-                },
-                "2-5-x": {
-                    "planned": "12/10/2024 10:51",
-                    "actual": null
-                },
-                "2-6-x": {
-                    "planned": "12/10/2024 10:51",
-                    "actual": null
-                }
-            }
-        }
+        ]
     },
     "tritium_detection": {
         "IV": {
@@ -306,6 +232,56 @@
                     "3": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv",
                     "4": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
                 }
+            },
+            "5": {
+                "planned_sample_time": "12/19/2024 10:51",
+                "actual_sample_time": null,
+                "lsc_vials_filenames": {
+                    "1": null,
+                    "2": null,
+                    "3": null,
+                    "4": null
+                }
+            },
+            "6": {
+                "planned_sample_time": "12/22/2024 10:51",
+                "actual_sample_time": null,
+                "lsc_vials_filenames": {
+                    "1": null,
+                    "2": null,
+                    "3": null,
+                    "4": null
+                }
+            },
+            "7": {
+                "planned_sample_time": "12/27/2024 10:51",
+                "actual_sample_time": null,
+                "lsc_vials_filenames": {
+                    "1": null,
+                    "2": null,
+                    "3": null,
+                    "4": null
+                }
+            },
+            "8": {
+                "planned_sample_time": "1/1/2025 10:51",
+                "actual_sample_time": null,
+                "lsc_vials_filenames": {
+                    "1": null,
+                    "2": null,
+                    "3": null,
+                    "4": null
+                }
+            },
+            "9": {
+                "planned_sample_time": "1/6/2025 10:51",
+                "actual_sample_time": null,
+                "lsc_vials_filenames": {
+                    "1": null,
+                    "2": null,
+                    "3": null,
+                    "4": null
+                }
             }
         },
         "OV": {
@@ -340,103 +316,5 @@
                 }
             }
         }
-    },
-    "lsc_samples": [
-        {
-            "label": "1L-OV_2-0-1",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-OV_2-0-2",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-OV_2-0-3",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-OV_2-0-4",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-0-1",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-0-2",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-0-3",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-0-4",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-1-1",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-1-2",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-1-3",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-1-4",
-            "filename": "1L_BL-2_IV-2-0_OV-2-0_IV-2-1_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-2-1",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-2-2",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-2-3",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-2-4",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-3-1",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-3-2",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-3-3",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-3-4",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-4-1",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-4-2",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-4-3",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        },
-        {
-            "label": "1L-IV_2-4-4",
-            "filename": "1L_BL-2_IV-2-3_IV-2-4_IV-2-2.csv"
-        }
-    ]
+    }
 }
